move luma.core.serial to luma.core.interface.serial
